+++ source/renderer/app/components/loading/syncing-connecting/SyncingConnecting.tsx
@@ -9,7 +9,10 @@ import SyncingConnectingStatus from './SyncingConnectingStatus';
 import { CardanoNodeStates } from '../../../../../common/types/cardano-node.types';
 // @ts-ignore ts-migrate(2307) FIXME: Cannot find module './SyncingConnecting.scss' or i... Remove this comment to see the full error message
 import styles from './SyncingConnecting.scss';
-import type { CardanoNodeState } from '../../../../../common/types/cardano-node.types';
+import type {
+  CardanoNodeState,
+  BlockSyncType,
+} from '../../../../../common/types/cardano-node.types';
 import { REPORT_ISSUE_TIME_TRIGGER } from '../../../config/timingConfig';
 import NewsFeedIcon from '../../widgets/NewsFeedIcon';
 
@@ -19,7 +22,10 @@ type State = {
 };
 type Props = {
   cardanoNodeState: CardanoNodeState | null | undefined;
-  verificationProgress: number;
+  blockSync: {
+    type: BlockSyncType;
+    progress: number;
+  };
   hasBeenConnected: boolean;
   forceConnectivityIssue?: boolean;
   isConnected: boolean;
@@ -164,7 +170,7 @@ class SyncingConnecting extends Component<Props, State> {
       onToggleNewsFeedIconClick,
       showNewsFeedIcon,
       isVerifyingBlockchain,
-      verificationProgress,
+      blockSync,
     } = this.props;
     const newsFeedIconStyles = classNames([
       isConnecting ? 'connectingScreen' : null,
@@ -205,7 +211,7 @@ class SyncingConnecting extends Component<Props, State> {
           isNodeStopping={isNodeStopping}
           isNodeStopped={isNodeStopped}
           isVerifyingBlockchain={isVerifyingBlockchain}
-          verificationProgress={verificationProgress}
+          blockSync={blockSync}
         />
         <StatusIcons
           onIconClick={onStatusIconClick}