+++ source/renderer/app/stores/ProfileStore.ts
@@ -9,7 +9,6 @@ import { ROUTES } from '../routes-config';
 import LocalizableError from '../i18n/LocalizableError';
 import { WalletSupportRequestLogsCompressError } from '../i18n/errors';
 import { generateFileNameWithTimestamp } from '../../../common/utils/files';
-import { formattedBytesToSize } from '../utils/formatters';
 import { logger } from '../utils/logging';
 import { setStateSnapshotLogChannel } from '../ipc/setStateSnapshotLogChannel';
 import { getDesktopDirectoryPathChannel } from '../ipc/getDesktopDirectoryPathChannel';
@@ -42,7 +41,7 @@ import {
   TIME_OPTIONS,
   PROFILE_SETTINGS,
 } from '../config/profileConfig';
-import formatCpuInfo from '../utils/formatCpuInfo';
+import { buildSystemInfo } from '../utils/buildSystemInfo';
 
 export default class ProfileStore extends Store {
   @observable
@@ -134,6 +133,8 @@ export default class ProfileStore extends Store {
   isSubmittingBugReport = false;
   @observable
   isInitialScreen = false;
+  @observable
+  isRTSModeRecommendationAcknowledged = false;
 
   /* eslint-enable max-len */
   setup() {
@@ -151,6 +152,9 @@ export default class ProfileStore extends Store {
     profileActions.getLogsAndCompress.listen(this._getLogsAndCompress);
     profileActions.downloadLogs.listen(this._downloadLogs);
     profileActions.downloadLogsSuccess.listen(this._toggleDisableDownloadLogs);
+    profileActions.acknowledgeRTSModeRecommendation.listen(
+      this._acknowledgeRTSFlagsModeRecommendation
+    );
     this.actions.app.initAppEnvironment.listen(() => {});
     this.registerReactions([
       this._updateBigNumberFormat,
@@ -369,6 +373,10 @@ export default class ProfileStore extends Store {
     // @ts-ignore ts-migrate(1320) FIXME: Type of 'await' operand must either be a valid pro... Remove this comment to see the full error message
     await this.getDataLayerMigrationAcceptanceRequest.execute();
   };
+  @action
+  _acknowledgeRTSFlagsModeRecommendation = () => {
+    this.isRTSModeRecommendationAcknowledged = true;
+  };
   _getDataLayerMigrationAcceptance = () => {
     this.getDataLayerMigrationAcceptanceRequest.execute();
   };
@@ -542,7 +550,6 @@ export default class ProfileStore extends Store {
         cardanoWalletPID,
         tlsConfig,
         stateDirectoryPath,
-        diskSpaceAvailable,
         cardanoNodeState,
         isConnected,
         isNodeInSync,
@@ -558,7 +565,6 @@ export default class ProfileStore extends Store {
         network,
         apiVersion,
         nodeVersion,
-        cpu,
         version,
         mainProcessID,
         rendererProcessID,
@@ -567,17 +573,8 @@ export default class ProfileStore extends Store {
         isMainnet,
         isStaging,
         isTestnet,
-        os,
-        platformVersion,
-        ram,
       } = this.environment;
-      const systemInfo = {
-        platform: os,
-        platformVersion,
-        cpu: formatCpuInfo(cpu),
-        ram: formattedBytesToSize(ram),
-        availableDiskSpace: diskSpaceAvailable,
-      };
+      const systemInfo = buildSystemInfo(this.environment, networkStatus);
       const coreInfo = {
         daedalusVersion: version,
         daedalusBuildNumber: build,
@@ -623,6 +620,7 @@ export default class ProfileStore extends Store {
       };
       await setStateSnapshotLogChannel.send(stateSnapshotData);
     } catch (error) {
+      // @ts-ignore ts-migrate(2554) FIXME: Expected 2 arguments, but got 1.
       logger.error('ProfileStore: State snapshot log file creation failed', {
         error,
       });